Adapt client to new Volume with mute flag and separate Volume component
