+++ components/restreamer/client/api/client.graphql
@@ -55,11 +55,17 @@ subscription State {
             dst
             label
             previewUrl
-            volume
+            volume {
+                level
+                muted
+            }
             mixins {
                 id
                 src
-                volume
+                volume {
+                    level
+                    muted
+                }
                 delay
             }
             enabled
@@ -180,13 +186,15 @@ mutation TuneVolume(
     $restream_id: RestreamId!
     $output_id: OutputId!
     $mixin_id: MixinId
-    $volume: Volume!
+    $level: VolumeLevel!
+    $muted: Boolean!
 ) {
     tuneVolume(
         restreamId: $restream_id
         outputId: $output_id
         mixinId: $mixin_id
-        volume: $volume
+        level: $level
+        muted: $muted
     )
 }
