+++ components/restreamer/client/src/components/Output.svelte
@@ -6,6 +6,7 @@
 
   import Confirm from './common/Confirm.svelte';
   import Toggle from './common/Toggle.svelte';
+  import Volume from './Volume.svelte';
   import Mixin from './Mixin.svelte';
   import RecordsModal from '../modals/RecordsModal.svelte';
   import Url from './common/Url.svelte';
@@ -30,23 +31,7 @@
 
   const tuneVolumeMutation = mutation(mutations.TuneVolume);
 
-  let volume = 100;
-  $: {
-    // Trigger Svelte reactivity watching.
-    value.volume = value.volume;
-    // Move `volume` to a separate function to omit triggering this block when
-    // it is changed, as we're only interested in `value` changes here.
-    update_volume();
-  }
-
   $: toggleStatusText = value.enabled ? 'Disable' : 'Enable';
-
-  // Last used non-zero volume.
-  let last_volume = value.volume === 0 ? 100 : value.volume;
-
-  function update_volume() {
-    volume = value.volume;
-  }
 
   async function toggle() {
     const variables = { restream_id, output_id: value.id };
@@ -68,23 +53,6 @@
     } catch (e) {
       showError(e.message);
     }
-  }
-
-  async function tuneVolume() {
-    if (volume !== 0) {
-      last_volume = volume;
-    }
-    const variables = { restream_id, output_id: value.id, volume };
-    try {
-      await tuneVolumeMutation({ variables });
-    } catch (e) {
-      showError(e.message);
-    }
-  }
-
-  async function toggleVolume() {
-    volume = volume !== 0 ? 0 : last_volume;
-    await tuneVolume();
   }
 
   function openEditOutputModal() {
@@ -203,7 +171,6 @@
         {/if}
       </div>
 
-      {#if value.mixins.length > 0}
         {#if !isMixPage()}
           <a
             class="single-view"
@@ -214,26 +181,8 @@
           </a>
         {/if}
 
-        <div class="uk-flex volume orig">
-          <a href="/" on:click|preventDefault={toggleVolume}>
-            {#if volume > 0}
-              <span><i class="fas fa-volume-up" title="Volume" /></span>
-            {:else}
-              <span><i class="fas fa-volume-mute" title="Muted" /></span>
-            {/if}
-          </a>
-          <input
-            class="uk-range"
-            type="range"
-            min="0"
-            max="200"
-            step="1"
-            bind:value={volume}
-            on:change={tuneVolume}
-          />
-          <span class="uk-margin-small-left">{volume}%</span>
-        </div>
-
+        <Volume volume={value.volume} restream_id={restream_id} output_id={value.id} {mutations}/>
+      {#if value.mixins.length > 0}
         {#each value.mixins as mixin}
           <Mixin {restream_id} output_id={value.id} value={mixin} {mutations} />
         {/each}
